feat(621): moved useQuery into separate file
