+++ packages/falcon-data/src/Query/Query.tsx
@@ -1,15 +1,11 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { DocumentNode } from 'graphql';
 import { OperationVariables, QueryResult } from '@apollo/react-common';
 import { Query as ApolloQuery, QueryComponentOptions } from '@apollo/react-components';
-import { useQuery as useApolloQuery, QueryHookOptions } from '@apollo/react-hooks';
 import { NetworkStatus } from 'apollo-client';
 import { Loader } from './Loader';
 import { OperationError } from './OperationError';
-
-export type ApolloFetchMore<TData, TVariables> = QueryResult<TData, TVariables>['fetchMore'];
-export type FetchMore<TData, TVariables> = (data: TData, fetchMore: ApolloFetchMore<TData, TVariables>) => any;
+import { FetchMore, ApolloFetchMore } from './fetchMore';
 
 export type QueryRenderProps<TData = any, TVariables = OperationVariables> = {
   fetchMore: (() => any) | ApolloFetchMore<TData, TVariables>;
@@ -58,21 +54,4 @@ export class Query<TData = any, TVariables = OperationVariables> extends React.C
       </ApolloQuery>
     );
   }
-}
-
-export type UseQueryOptions<TData, TVariables> = QueryHookOptions<TData, TVariables> & {
-  fetchMore?: FetchMore<TData, TVariables>;
-};
-export function useQuery<TData = any, TVariables = OperationVariables>(
-  query: DocumentNode,
-  options?: UseQueryOptions<TData, TVariables>
-) {
-  const { fetchMore, ...restOptions } = options || ({} as UseQueryOptions<TData, TVariables>);
-  const result = useApolloQuery<TData, TVariables>(query, restOptions);
-  const { data, fetchMore: apolloFetchMore } = result;
-
-  return {
-    ...result,
-    fetchMore: fetchMore ? () => fetchMore(data!, result.fetchMore) : apolloFetchMore
-  };
 }
