+++ packages/falcon-data/src/Query/index.ts
@@ -1,4 +1,6 @@
 export * from './Query';
+export * from './useQuery';
+export * from './fetchMore';
 export * from './Loader';
 export * from './apolloErrorToErrorModelList';
 export * from './OperationError';
